backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
neo4j==5.12.0
pdfminer.six==20221105
spacy==3.7.0
transformers==4.35.0
torch==2.2.0
pydantic==1.10.12
python-dotenv==1.0.0
